Properly integrated knoboo with twistd. Knoboo is now a set of twistd plugins:  - knoboo  - knobood  - knoboo-kernel (run just twistd in knoboo's project root to see these plugins listed)
config.py is no longer needed and has been deleted.
main.py is no longer needed, but hasn't been deleted yet.

Remaining issues:
There is still a dependency on ~/.knoboo; need to fix that to point to Knoboo/data
There are commented hacks in service.py; the next iteration should focus on resolving these

The settings for each service should be clearly organized, and possibly split into separate files.

The distutils packaging has not been updated, so knoboo should not be installed with setup.py.
For now it must be run with twistd in the root knoboo project directory.